commit for modified and testing app

+++ src/Controller/TaskController.php
@@ -20,6 +20,22 @@ class TaskController extends AbstractController
     #[Route('/', name: 'task_list')]
     public function index(TaskService $taskService): Response
     {
+        $user = $this->getUser();
+        if ($user === null) {
+            
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
+        return $this->render('Tasks/list.html.twig',['tasks'=> $taskService->getPaginatedTasks()]);
+    }
+
+    #[Route('/finished', name: 'task_list_finished')]
+    public function finishedTaskList(TaskService $taskService): Response
+    {
+        $user = $this->getUser();
+        if ($user === null) {
+            
+            return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
+        }
         return $this->render('Tasks/list.html.twig',['tasks'=> $taskService->getPaginatedTasks()]);
     }
 
@@ -38,7 +54,6 @@ class TaskController extends AbstractController
 
         if ($user === null) {
             
-            $this->addFlash('danger', 'Veuillez vous connecter pour ajouter un task');
             return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
         }
         //I check if I have a form and that it is valid
@@ -70,7 +85,7 @@ class TaskController extends AbstractController
     {
         $task = new Task;
         $taskId = $request->attributes->get('id');
-        $task = $taskRepository->findTaskWithUserAndCategory($taskId);
+        $task = $taskRepository->findTaskWithUser($taskId);
 
 
         // $this->denyAccessUnlessGranted('task_EDIT', $task);
@@ -82,7 +97,6 @@ class TaskController extends AbstractController
 
         if ($user === null) {
             
-            $this->addFlash('danger', 'Veuillez vous connecter pour modifier une tâche');
             return $this->redirectToRoute('app_login', [], Response::HTTP_SEE_OTHER);
         }
         //I check if I have a form and that it is valid
@@ -110,14 +124,19 @@ class TaskController extends AbstractController
         ]);
     }
     
-    #[Route('/{id}/toggle', name: 'task_toggle')]
+    #[Route('/toggle/{id}', name: 'task_toggle')]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em)
     {
-        $task->setIsDone(!$task->isIsDone());
-        $em->flush($task);
-
-        $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
-
+        if ($this->getUser()->getId() === $task->getUser()->getId() || $this->getUser()->getRoles()[0] == "ROLE_ADMIN") {
+            $task->setIsDone(!$task->isIsDone());
+            $em->flush($task);
+    
+            $this->addFlash('success', sprintf('La tâche %s a bien été marquée comme faite.', $task->getTitle()));
+    
+            return $this->redirectToRoute('task_list');
+        }
+        $this->addFlash('danger', sprintf('La tâche %s n\'a pas été marquée comme faite.', $task->getTitle()));
+    
         return $this->redirectToRoute('task_list');
     }
 
@@ -125,8 +144,12 @@ class TaskController extends AbstractController
     public function delete( Task $task, TaskRepository $taskRepository): Response
     {
         // $this->denyAccessUnlessGranted('TRICK_DELETE', $trick);
-        $taskRepository->remove($task, true);
-        $this->addFlash('success', sprintf('La tâche a bien eté supprimée'));
+        
+        if ($this->getUser()->getId() === $task->getUser()->getId() || $this->getUser()->getRoles()[0] == "ROLE_ADMIN") {
+            $taskRepository->remove($task, true);
+            $this->addFlash('success', sprintf('La tâche a bien eté supprimée'));
+        }
+
         return $this->redirectToRoute('task_list', [], Response::HTTP_SEE_OTHER);
     }
 }
